Move crafting related functions to character
Shake the tree of moving crafting functions to Character, and pick up
whatever falls out.
