+++ src/game_inventory.h
@@ -86,7 +86,7 @@ item_location consume_meds( player &p );
 /** Choosing a container for liquid. */
 item_location container_for( avatar &you, const item &liquid, int radius = 0 );
 /** Item disassembling menu. */
-item_location disassemble( player &p );
+item_location disassemble( Character &p );
 /** Gunmod installation menu. */
 item_location gun_to_modify( player &p, const item &gunmod );
 /** Book reading menu. */
